Servo aansturing SPRINT1
This firmware contains 2 functions and 1 struct.

The user can insert the servo positions through the serial monitor (input) in this format: s0s1s2 > 180180180 for all to go to 180 degrees.

The posistions are saved in the struct with will be used in the servo function with will be called at the end of the loop

https://github.com/EHammer98/DigitalPlastic/issues/1